Use default linker flags with CMake build
Already these flags are included by the new and legacy toolchains. It is
possible that by excluding other linker flags set by these toolchains;
the macOS build is failing. So go ahead and drop the linker flag
customization for CMake and just use the new and legacy toolchains'
flags as is.